initial upload of forked files with tweaked readme
Hi folks. I hope this is helpful. I really love how fun and feature-packed the MC101 is and I realize that it is not the most intuitive device. The main reason I forked the cheatsheet is that the previous repo appeared to no longer being maintained, and I didn't want to clutter up someone else's life with updated info/tweaks. (I guess I'm shy like that).

Anyway, the biggest change is that the 101 now has an actual **Undo**/**redo** feature added. In addition to a host of new ways to tweak synthesizer voices.
Version 1.82 of the firmware, released in June, 2023, includes stability improvements which make a world of difference for loading zen core sound packs (`.SDZ` files).